new: Add Favorites btn toggle
