+++ src/components/places-container/places-container.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import PlacesList from '../places-list/places-list';
+import Sort from '../sort/sort';
 import MapSection from '../map-section/map-section';
-const MAIN_PLACE_CARD_COUNT = 4;
+import {CountCards} from '../../utils/const';
 
 const PlacesContainer = ({filteredOffers, activeCity, onCardClick}) => {
   return (
@@ -10,22 +11,8 @@ const PlacesContainer = ({filteredOffers, activeCity, onCardClick}) => {
       <section className="cities__places places">
         <h2 className="visually-hidden">Places</h2>
         <b className="places__found">{filteredOffers.length} {filteredOffers.length === 1 ? `place` : `places`} to stay in {activeCity}</b>
-        <form className="places__sorting" action="#" method="get">
-          <span className="places__sorting-caption">Sort by</span>
-          <span className="places__sorting-type" tabIndex="0">
-  Popular
-            <svg className="places__sorting-arrow" width="7" height="4">
-              <use xlinkHref="#icon-arrow-select"></use>
-            </svg>
-          </span>
-          <ul className="places__options places__options--custom places__options--opened">
-            <li className="places__option places__option--active" tabIndex="0">Popular</li>
-            <li className="places__option" tabIndex="0">Price: low to high</li>
-            <li className="places__option" tabIndex="0">Price: high to low</li>
-            <li className="places__option" tabIndex="0">Top rated first</li>
-          </ul>
-        </form>
-        <PlacesList cardListClass={`cities__places-list tabs__content`} containerCardClass={`cities__place-card`} imageCardClass={`cities__image-wrapper`} offers={filteredOffers} onCardClick={onCardClick} cardsCount={MAIN_PLACE_CARD_COUNT}/>
+        <Sort />
+        <PlacesList cardListClass={`cities__places-list tabs__content`} containerCardClass={`cities__place-card`} imageCardClass={`cities__image-wrapper`} offers={filteredOffers} onCardClick={onCardClick} cardsCount={CountCards.MAIN_LIST}/>
       </section>
       <div className="cities__right-section">
         <MapSection sectionName={`cities`}/>
